backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
motor
openai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-dotenv